requirements.txt
django
gunicorn
psycopg2
django-debug-toolbar
django-mptt
linaro-django-pagination
django-filter
webstack-django-sorting
djangorestframework
Pillow
django-haystack
pyelasticsearch
django-froala-editor